@@ -26,21 +26,22 @@
     "@types/express-session": "^1.15.0",
     "@types/helmet": "0.0.35",
     "@types/isomorphic-fetch": "0.0.34",
-    "@types/leaflet": "^1.0.65",
-    "@types/lodash": "^4.14.68",
+    "@types/leaflet": "^1.2.9",
+    "@types/lodash": "^4.14.116",
     "@types/normalizr": "^2.0.18",
-    "@types/prop-types": "^15.5.1",
-    "@types/react": "^15.0.34",
+    "@types/prop-types": "^15.5.5",
+    "@types/react": "^16.4.11",
     "@types/react-addons-css-transition-group": "^15.0.5",
-    "@types/react-dom": "^15.5.1",
+    "@types/react-dom": "^16.0.7",
     "@types/react-intl": "^2.3.1",
-    "@types/react-leaflet": "^1.1.3",
+    "@types/react-leaflet": "^1.1.6",
     "@types/react-redux": "5.0.1",
     "@types/react-router": "^4.0.12",
     "@types/react-router-config": "^1.0.2",
     "@types/react-router-dom": "^4.0.5",
     "@types/react-router-redux": "^5.0.3",
     "@types/react-sidebar": "2.3.0",
+    "@types/react-transition-group": "^1.1.5",
     "@types/redux": "^3.6.0",
     "@types/redux-actions": "^1.2.6",
     "@types/redux-logger": "^3.0.0",
@@ -72,19 +73,19 @@
     "lodash": "^4.17.10",
     "normalizr": "^3.2.3",
     "pg": "^6.4.0",
-    "prop-types": "^15.5.10",
-    "react": "^15.6.1",
-    "react-addons-css-transition-group": "^15.6.2",
-    "react-dom": "^15.6.1",
+    "prop-types": "^15.6.2",
+    "react": "^16.4.2",
+    "react-dom": "^16.4.2",
     "react-intl": "^2.3.0",
     "react-leaflet": "^1.7.1",
-    "react-leaflet-div-icon": "^1.0.2",
+    "react-leaflet-div-icon": "^1.1.0",
     "react-redux": "^5.0.5",
-    "react-router": "^4.1.1",
+    "react-router": "^4.2.0",
     "react-router-config": "^1.0.0-beta.3",
     "react-router-dom": "^4.1.1",
     "react-router-redux": "^4.0.8",
     "react-sidebar": "2.3.2",
+    "react-transition-group": "^1.2.1",
     "redux": "^3.7.1",
     "redux-actions": "^2.2.1",
     "redux-logger": "^3.0.6",
@@ -121,7 +122,7 @@
     "style-loader": "^0.18.2",
     "ts-loader": "^2.2.2",
     "ts-node": "^3.2.0",
-    "typescript": "^2.4.1",
+    "typescript": "^2.9.2",
     "url-loader": "^0.5.9",
     "webpack": "^3.1.0",
     "webpack-dashboard": "^0.4.0",
